+++ lib/src/form/smart_field_registry.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: public_member_api_docs
+
 import 'dart:ui' show Offset;
 
 import 'package:flutter/foundation.dart';