+++ lib/src/form/smart_form_scope.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: public_member_api_docs
+
 import 'package:flutter/widgets.dart';
 
 import '../animation/smart_error_animation.dart';